Refactor structs; Reorganize files
