+++ crates/foundry-ui/src/message.rs
@@ -1,24 +1,21 @@
 use serde::Serialize;
 
-use crate::formats::NumbersFormat;
-
 pub trait Message {
     /// Return textual representation of this message.
     ///
     /// Default implementation returns empty string, making [`Ui`] skip printing this message.
-    fn text(&self, numbers_format: NumbersFormat) -> String
+    fn text(&self) -> String
     where
         Self: Sized + Serialize,
     {
-        let _ = numbers_format;
         String::new()
     }
 
-    fn print_human(&self, numbers_format: NumbersFormat, print_as_err: bool)
+    fn print_human(&self, print_as_err: bool)
     where
         Self: Sized + Serialize,
     {
-        let text = self.text(numbers_format);
+        let text = self.text();
         if !text.is_empty() && print_as_err {
             eprintln!("{text}");
         } else if !text.is_empty() && !print_as_err {
@@ -26,19 +23,18 @@ pub trait Message {
         }
     }
 
-    fn json(&self, numbers_format: NumbersFormat) -> String
+    fn json(&self) -> String
     where
         Self: Sized + Serialize,
     {
-        let _ = numbers_format;
         serde_json::to_string(self).unwrap_or_else(|_| "Invalid JSON".to_string())
     }
 
-    fn print_json(&self, numbers_format: NumbersFormat, print_as_err: bool)
+    fn print_json(&self, print_as_err: bool)
     where
         Self: Serialize + Sized,
     {
-        let json = self.json(numbers_format);
+        let json = self.json();
         if print_as_err {
             eprintln!("{json}");
         } else {
@@ -48,15 +44,13 @@ pub trait Message {
 }
 
 impl Message for &str {
-    fn text(&self, numbers_format: NumbersFormat) -> String {
-        let _ = numbers_format;
+    fn text(&self) -> String {
         (*self).to_string()
     }
 }
 
 impl Message for String {
-    fn text(&self, numbers_format: NumbersFormat) -> String {
-        let _ = numbers_format;
+    fn text(&self) -> String {
         self.to_string()
     }
 }
