+++ crates/sncast/src/response/account/create.rs
@@ -1,0 +1,71 @@
+use conversions::padded_felt::PaddedFelt;
+use foundry_ui::{Message, formats::OutputFormat};
+use serde::{Serialize, Serializer};
+use starknet_types_core::felt::Felt;
+
+use crate::{
+    helpers::block_explorer::LinkProvider,
+    response::{
+        cast_message::CastMessage,
+        command::CommandResponse,
+        explorer_link::OutputLink,
+        print::{Format, OutputData},
+    },
+};
+pub struct Decimal(pub u64);
+
+impl Serialize for Decimal {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(&self.0.to_string())
+    }
+}
+
+fn serialize_as_decimal<S>(value: &Felt, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&format!("{value:#}"))
+}
+
+#[derive(Serialize, Debug)]
+pub struct AccountCreateResponse {
+    pub address: PaddedFelt,
+    #[serde(serialize_with = "serialize_as_decimal")]
+    pub max_fee: Felt,
+    pub add_profile: String,
+    pub message: String,
+}
+
+impl CommandResponse for AccountCreateResponse {}
+
+impl Message for AccountCreateResponse {}
+
+impl CastMessage<AccountCreateResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("account create", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("account create", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
+
+impl OutputLink for AccountCreateResponse {
+    const TITLE: &'static str = "account creation";
+
+    fn format_links(&self, provider: Box<dyn LinkProvider>) -> String {
+        format!("account: {}", provider.contract(self.address))
+    }
+}
