+++ crates/sncast/src/response/account/delete.rs
@@ -1,0 +1,36 @@
+use foundry_ui::{Message, formats::OutputFormat};
+use serde::Serialize;
+
+use crate::response::{
+    cast_message::CastMessage,
+    command::CommandResponse,
+    print::{Format, OutputData},
+};
+
+#[derive(Serialize)]
+pub struct AccountDeleteResponse {
+    pub result: String,
+}
+
+impl CommandResponse for AccountDeleteResponse {}
+
+impl Message for AccountDeleteResponse {}
+
+impl CastMessage<AccountDeleteResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("account delete", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("account delete", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
