+++ crates/sncast/src/response/cast_message.rs
@@ -1,0 +1,6 @@
+use foundry_ui::{Message, formats::NumbersFormat};
+
+pub struct CastMessage<T: Message> {
+    pub numbers_format: NumbersFormat,
+    pub message: T,
+}
