+++ crates/sncast/src/response/command.rs
@@ -1,0 +1,3 @@
+use serde::Serialize;
+
+pub trait CommandResponse: Serialize {}
