+++ crates/sncast/src/response/declare.rs
@@ -1,0 +1,95 @@
+use conversions::{padded_felt::PaddedFelt, serde::serialize::CairoSerialize};
+use foundry_ui::{Message, formats::OutputFormat};
+use indoc::formatdoc;
+use serde::{Deserialize, Serialize};
+
+use crate::helpers::block_explorer::LinkProvider;
+
+use super::{
+    cast_message::CastMessage,
+    command::CommandResponse,
+    explorer_link::OutputLink,
+    print::{Format, OutputData},
+};
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+pub struct DeclareTransactionResponse {
+    pub class_hash: PaddedFelt,
+    pub transaction_hash: PaddedFelt,
+}
+
+impl CommandResponse for DeclareTransactionResponse {}
+
+impl Message for DeclareTransactionResponse {}
+
+impl CastMessage<DeclareTransactionResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("declare", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("declare", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+pub struct AlreadyDeclaredResponse {
+    pub class_hash: PaddedFelt,
+}
+
+impl CommandResponse for AlreadyDeclaredResponse {}
+
+impl Message for AlreadyDeclaredResponse {}
+
+impl CastMessage<AlreadyDeclaredResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("declare", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("declare", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
+
+#[derive(Clone, Serialize, Deserialize, CairoSerialize, Debug, PartialEq)]
+#[serde(tag = "status")]
+pub enum DeclareResponse {
+    AlreadyDeclared(AlreadyDeclaredResponse),
+    #[serde(untagged)]
+    Success(DeclareTransactionResponse),
+}
+
+impl CommandResponse for DeclareResponse {}
+
+impl OutputLink for DeclareTransactionResponse {
+    const TITLE: &'static str = "declaration";
+
+    fn format_links(&self, provider: Box<dyn LinkProvider>) -> String {
+        formatdoc!(
+            "
+            class: {}
+            transaction: {}
+            ",
+            provider.class(self.class_hash),
+            provider.transaction(self.transaction_hash)
+        )
+    }
+}
