+++ crates/sncast/src/response/invoke.rs
@@ -1,0 +1,50 @@
+use crate::helpers::block_explorer::LinkProvider;
+
+use super::{
+    cast_message::CastMessage,
+    command::CommandResponse,
+    explorer_link::OutputLink,
+    print::{Format, OutputData},
+};
+use conversions::{padded_felt::PaddedFelt, serde::serialize::CairoSerialize};
+use foundry_ui::{Message, formats::OutputFormat};
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, CairoSerialize, Clone, Debug, PartialEq)]
+pub struct InvokeResponse {
+    pub transaction_hash: PaddedFelt,
+}
+
+impl CommandResponse for InvokeResponse {}
+
+impl Message for InvokeResponse {}
+
+impl CastMessage<InvokeResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("invoke", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("invoke", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
+
+impl OutputLink for InvokeResponse {
+    const TITLE: &'static str = "invocation";
+
+    fn format_links(&self, provider: Box<dyn LinkProvider>) -> String {
+        format!(
+            "transaction: {}",
+            provider.transaction(self.transaction_hash)
+        )
+    }
+}
