+++ crates/sncast/src/response/script/mod.rs
@@ -1,0 +1,2 @@
+pub mod init;
+pub mod run;
