+++ crates/sncast/src/response/tx_status.rs
@@ -1,0 +1,52 @@
+use conversions::serde::serialize::CairoSerialize;
+use foundry_ui::{Message, formats::OutputFormat};
+use serde::Serialize;
+
+use super::{
+    cast_message::CastMessage,
+    command::CommandResponse,
+    print::{Format, OutputData},
+};
+
+#[derive(Serialize, CairoSerialize)]
+pub enum FinalityStatus {
+    Received,
+    Rejected,
+    AcceptedOnL2,
+    AcceptedOnL1,
+}
+
+#[derive(Serialize, CairoSerialize)]
+pub enum ExecutionStatus {
+    Succeeded,
+    Reverted,
+}
+
+#[derive(Serialize, CairoSerialize)]
+pub struct TransactionStatusResponse {
+    pub finality_status: FinalityStatus,
+    pub execution_status: Option<ExecutionStatus>,
+}
+
+impl CommandResponse for TransactionStatusResponse {}
+
+impl Message for TransactionStatusResponse {}
+
+impl CastMessage<TransactionStatusResponse> {
+    // TODO(#3391): Update text output to be more user friendly
+    #[must_use]
+    pub fn text(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("tx-status", OutputFormat::Human)
+            .expect("Failed to format response")
+    }
+
+    #[must_use]
+    pub fn json(&self) -> String {
+        OutputData::from(&self.message)
+            .format_with(self.numbers_format)
+            .to_string_pretty("tx-status", OutputFormat::Json)
+            .expect("Failed to format response")
+    }
+}
