+++ crates/sncast/src/starknet_commands/account/create.rs
@@ -13,7 +13,7 @@ use sncast::helpers::constants::{
     CREATE_KEYSTORE_PASSWORD_ENV_VAR, OZ_CLASS_HASH,
 };
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::structs::AccountCreateResponse;
+use sncast::response::account::create::AccountCreateResponse;
 use sncast::{
     AccountType, Network, check_class_hash_exists, check_if_legacy_contract,
     extract_or_generate_salt, get_chain_id, get_keystore_password, handle_account_factory_error,
