+++ crates/sncast/src/starknet_commands/account/import.rs
@@ -13,7 +13,7 @@ use sncast::helpers::account::generate_account_name;
 use sncast::helpers::braavos::assert_non_braavos_account;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::structs::AccountImportResponse;
+use sncast::response::account::import::AccountImportResponse;
 use sncast::{AccountType, check_class_hash_exists, get_chain_id, handle_rpc_error};
 use starknet::core::types::{BlockId, BlockTag, StarknetError};
 use starknet::providers::jsonrpc::{HttpTransport, JsonRpcClient};
