+++ crates/sncast/src/starknet_commands/call.rs
@@ -2,8 +2,8 @@ use crate::Arguments;
 use anyhow::Result;
 use clap::Args;
 use sncast::helpers::rpc::RpcArgs;
+use sncast::response::call::CallResponse;
 use sncast::response::errors::StarknetCommandError;
-use sncast::response::structs::CallResponse;
 use starknet::core::types::{BlockId, FunctionCall};
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::{JsonRpcClient, Provider};
