+++ crates/sncast/src/starknet_commands/declare.rs
@@ -5,10 +5,10 @@ use conversions::byte_array::ByteArray;
 use scarb_api::StarknetContractArtifacts;
 use sncast::helpers::fee::{FeeArgs, FeeSettings};
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::errors::StarknetCommandError;
-use sncast::response::structs::{
+use sncast::response::declare::{
     AlreadyDeclaredResponse, DeclareResponse, DeclareTransactionResponse,
 };
+use sncast::response::errors::StarknetCommandError;
 use sncast::{ErrorData, WaitForTx, apply_optional_fields, handle_wait_for_tx};
 use starknet::accounts::AccountError::Provider;
 use starknet::accounts::{ConnectedAccount, DeclarationV3};
