+++ crates/sncast/src/starknet_commands/deploy.rs
@@ -3,8 +3,8 @@ use clap::Args;
 use conversions::IntoConv;
 use sncast::helpers::fee::{FeeArgs, FeeSettings};
 use sncast::helpers::rpc::RpcArgs;
+use sncast::response::deploy::DeployResponse;
 use sncast::response::errors::StarknetCommandError;
-use sncast::response::structs::DeployResponse;
 use sncast::{WaitForTx, apply_optional_fields, handle_wait_for_tx};
 use sncast::{extract_or_generate_salt, udc_uniqueness};
 use starknet::accounts::AccountError::Provider;
