+++ crates/sncast/src/starknet_commands/invoke.rs
@@ -5,7 +5,7 @@ use conversions::IntoConv;
 use sncast::helpers::fee::{FeeArgs, FeeSettings};
 use sncast::helpers::rpc::RpcArgs;
 use sncast::response::errors::StarknetCommandError;
-use sncast::response::structs::InvokeResponse;
+use sncast::response::invoke::InvokeResponse;
 use sncast::{WaitForTx, apply_optional_fields, handle_wait_for_tx};
 use starknet::accounts::AccountError::Provider;
 use starknet::accounts::{Account, ConnectedAccount, ExecutionV3, SingleOwnerAccount};
