+++ crates/sncast/src/starknet_commands/show_config.rs
@@ -3,7 +3,8 @@ use camino::Utf8PathBuf;
 use clap::Args;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::structs::{Decimal, ShowConfigResponse};
+use sncast::response::account::create::Decimal;
+use sncast::response::show_config::ShowConfigResponse;
 use sncast::{chain_id_to_network_name, get_chain_id};
 use starknet::providers::JsonRpcClient;
 use starknet::providers::jsonrpc::HttpTransport;
